backend/requirements.txt
requests
torch
numpy
flask
gtts
pydub
speechrecognition
python-dotenv
openai
whisper
langchain-openai
langchain-community
pypdf
docarray
PyAudio
beautifulsoup4
youtube-transcript-api